feat: add interface to support copy in batches
